+++ es-api-model/src/main/kotlin/tech/figure/eventstream/stream/models/TxEvent.kt
@@ -1,6 +1,7 @@
 package tech.figure.eventstream.stream.models
 
 import com.squareup.moshi.JsonClass
+import java.math.BigInteger
 import java.time.OffsetDateTime
 
 /**
@@ -14,7 +15,7 @@ data class TxEvent(
     val txHash: String,
     override val eventType: String,
     override val attributes: List<Event>,
-    val fee: Long?,
+    val fee: BigInteger?,
     val denom: String?,
     val note: String?
 ) : EncodedBlockchainEvent
